LayoutTests/accessibility/presentational-elements-with-focus.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="group" tabindex="0" id="group">
<img tabindex="0" role="presentation" src="missing.gif" height="10" width="100" aria-label="foo" alt="bar" title="baz"><br><br>
<img tabindex="0" role="presentation" src="missing.gif" height="10" width="100" alt="foo1" title="bar"><br><br>
<img tabindex="0" role="presentation" src="missing.gif" height="10" width="100" alt="" title="foo"><br><br>
<img tabindex="0" role="presentation" src="missing.gif" height="10" width="100" title="foo"><br><br>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the ARIA presentation role is ignored when an element can set focus.");

    if (window.accessibilityController) {

          document.getElementById("group").focus();

          // Verify that all four images are accessible.
          shouldBe("accessibilityController.focusedElement.childrenCount", "4");
    }

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
